checkpoint, mostly working
