+++ Sources/YouVersionPlatformReader/Sheets/BibleReaderMyVersionsView.swift
@@ -3,7 +3,7 @@ import YouVersionPlatformCore
 import YouVersionPlatformUI
 
 public struct BibleReaderMyVersionsView: View {
-    @Environment(BibleReaderViewModel.self) private var viewModel
+    @Environment(BibleVersionsViewModel.self) private var viewModel
 
     public var body: some View {
         VStack {
