+++ Sources/YouVersionPlatformReader/Sheets/BibleReaderVersionListView.swift
@@ -3,7 +3,7 @@ import YouVersionPlatformCore
 import YouVersionPlatformUI
 
 public struct BibleReaderVersionListView: View {
-    @Environment(BibleReaderViewModel.self) private var viewModel
+    @Environment(BibleVersionsViewModel.self) private var viewModel
     @State private var searchText = ""
 
     public var body: some View {
@@ -135,5 +135,5 @@ public struct BibleReaderVersionListView: View {
 
 #Preview {
     BibleReaderVersionListView()
-        .environment(BibleReaderViewModel.preview)
+        .environment(BibleVersionsViewModel.preview)
 }
