+++ Sources/YouVersionPlatformReader/Sheets/BibleReaderVersionsStack.swift
@@ -2,14 +2,14 @@ import SwiftUI
 import YouVersionPlatformCore
 
 struct BibleReaderVersionsStack: View {
-    @Environment(BibleReaderViewModel.self) private var viewModel
+    @Environment(BibleVersionsViewModel.self) private var viewModel
 
     var body: some View {
         @Bindable var bindableViewModel = viewModel
 
         NavigationStack(path: $bindableViewModel.versionsPickerStack) {
             rootView
-                .navigationDestination(for: BibleReaderViewModel.VersionsPickerScreen.self) { screen in
+                .navigationDestination(for: BibleVersionsViewModel.VersionsPickerScreen.self) { screen in
                     destinationView(for: screen)
                 }
         }
@@ -34,7 +34,7 @@ struct BibleReaderVersionsStack: View {
     }
 
     @ViewBuilder
-    private func destinationView(for screen: BibleReaderViewModel.VersionsPickerScreen) -> some View {
+    private func destinationView(for screen: BibleVersionsViewModel.VersionsPickerScreen) -> some View {
         switch screen {
         case .myVersions:
             BibleReaderMyVersionsView()
@@ -45,7 +45,7 @@ struct BibleReaderVersionsStack: View {
         case .versionDownload:
             BibleVersionDownloadView()
         case .languages:
-            BibleReaderLanguagesView(viewModel: viewModel)
+            BibleReaderLanguagesView()
         }
     }
 }
